% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.ci-package.R
\docType{data}
\name{critical.value.hall.95}
\alias{critical.value.hall.95}
\title{Critical Values}
\source{
Klein, Moeschberger (2002): Survival Analysis, Springer
}
\description{
Critical values for the 95 \% Hall-Wellner band.
}
\details{
These values are taken from the book by Klein & Moeschberger.
}
\keyword{datasets}
