\name{critical.value.hall.99}
\alias{critical.value.hall.99}
\docType{data}
\title{Critical Values}
\description{
Critical values for the 99 \% Hall-Wellner band.
}
\usage{data(critical.value.hall.99)}
\details{
  These values are taken from the book by Klein & Moeschberger.
}
\source{
Klein, Moeschberger (2002): Survival Analysis, Springer.
}
\keyword{datasets}
