% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.ci-package.R
\docType{data}
\name{critical.value.nair.90}
\alias{critical.value.nair.90}
\title{Critical Values}
\source{
Klein, Moeschberger (2002): Survival Analysis, Springer.
}
\description{
Critical values for the 90 \% equal precision band by Nair.
}
\details{
These values are taken from the book by Klein & Moeschberger.
}
\keyword{datasets}
