% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.ci-package.R
\docType{data}
\name{rectum.dat}
\alias{rectum.dat}
\title{Rectum carcinoma data set.}
\format{
A data frame with 205 observations on the following 2 variables.
\describe{ \item{time}{Time in months} \item{status}{Status
at dropout} }
}
\source{
Merkel, Mansmann et al.(2001).The prognostic inhomogeneity in pT3
rectal carcinomas.  Int J Colorectal Dis.16, 305--306.
}
\description{
The rectum data contains 205 persons from a study about the survival of
patients with rectum carcinoma. Due to the severe course of disease the
follow-up was almost perfect in these data and involves hardly any censoring
and survivors. The data was used to analyze the behavior of the confidence
intervals in data sets with low censoring rate.
}
\keyword{datasets}
