\name{critical.value.nair.90}
\alias{critical.value.nair.90}
\docType{data}
\title{Critical Values}
\description{
Critical values for the 90 \% equal precision band by Nair.
}
\usage{data(critical.value.hall.90)}
\details{
  These values are taken from the book by Klein & Moeschberger.
}
\source{
Klein, Moeschberger (2002): Survival Analysis, Springer.
}
\keyword{datasets}
